refactor(bridge): extract message content census
